+++ packages/dso-toolkit/src/components/segmented-button/index.ts
@@ -1,0 +1,3 @@
+export * from "./segmented-button.args.js";
+export * from "./segmented-button.models.js";
+export * from "./segmented-button.stories-of.js";