+++ packages/dso-toolkit/src/components/segmented-button/segmented-button.args.ts
@@ -1,0 +1,70 @@
+import { HandlerFunction } from "storybook/actions";
+import { ArgTypes } from "storybook/internal/types";
+import { fn } from "storybook/test";
+
+import { argTypeAction } from "../../storybook";
+
+import { SegmentedButton } from "./segmented-button.models.js";
+
+export interface SegmentedButtonArgs {
+  options: SegmentedButton["options"];
+  activeOption?: number;
+  dsoChange: HandlerFunction;
+  segmentedAriaRequired?: boolean;
+  segmentedAriaLabel?: string;
+}
+
+export const segmentedButtonArgs: SegmentedButtonArgs = {
+  options: [
+    {
+      label: "Button 1",
+    },
+    {
+      label: "Button 2",
+    },
+    {
+      label: "Button 3",
+    },
+    {
+      label: "Button 4",
+    },
+  ],
+  activeOption: 0,
+  segmentedAriaRequired: true,
+  segmentedAriaLabel: "Segmented button options",
+  dsoChange: fn(),
+};
+
+export const segmentedButtonArgTypes: ArgTypes<SegmentedButtonArgs> = {
+  options: {
+    control: {
+      type: "object",
+    },
+  },
+  activeOption: {
+    options: [undefined, 0, 1, 2, 3, 4],
+    control: {
+      type: "select",
+    },
+  },
+  dsoChange: argTypeAction(),
+  segmentedAriaRequired: {
+    control: {
+      type: "boolean",
+    },
+  },
+  segmentedAriaLabel: {
+    control: {
+      type: "text",
+    },
+  },
+};
+
+export function segmentedButtonArgsMapper(a: SegmentedButtonArgs): SegmentedButton {
+  return {
+    ...a,
+    dsoChange: (event) => {
+      a.dsoChange(event.detail);
+    },
+  };
+}