+++ packages/dso-toolkit/src/components/segmented-button/segmented-button.models.ts
@@ -1,0 +1,17 @@
+export interface SegmentedButtonChangeEvent {
+  originalEvent: Event;
+  option: SegmentedButtonOption;
+}
+
+export interface SegmentedButton {
+  options: SegmentedButtonOption[];
+  activeOption?: number;
+  dsoChange?: (event: CustomEvent<SegmentedButtonChangeEvent>) => void;
+  segmentedAriaRequired?: boolean;
+  segmentedAriaLabel?: string;
+}
+
+export interface SegmentedButtonOption {
+  label: string;
+  disabled?: boolean;
+}