+++ storybook/src/components/segmented-button/segmented-button.core-template.ts
@@ -1,0 +1,29 @@
+import { DsoSegmentedButtonCustomEvent, SegmentedButtonChangeEvent } from "@dso-toolkit/core";
+import { html } from "lit-html";
+import { ifDefined } from "lit-html/directives/if-defined.js";
+
+import type { SegmentedButton } from "../../../../packages/dso-toolkit/src/components/segmented-button/segmented-button.models";
+import { ComponentImplementation } from "../../templates";
+
+export const coreSegmentedButton: ComponentImplementation<SegmentedButton> = {
+  component: "segmentedButton",
+  implementation: "core",
+  template: () =>
+    function segmentedButtonTemplate({ options, activeOption, dsoChange, segmentedAriaRequired, segmentedAriaLabel }) {
+      function statefulChangeHandler(event: DsoSegmentedButtonCustomEvent<SegmentedButtonChangeEvent>) {
+        dsoChange?.(event);
+
+        event.target.activeOption = event.detail.option;
+      }
+
+      return html`
+        <dso-segmented-button
+          .options=${options}
+          active-option=${ifDefined(activeOption)}
+          ?segmented-aria-required=${segmentedAriaRequired}
+          segmented-aria-label=${ifDefined(segmentedAriaLabel)}
+          @dsoChange=${statefulChangeHandler}
+        ></dso-segmented-button>
+      `;
+    },
+};