+++ storybook/src/templates.ts
@@ -87,6 +87,7 @@ import {
   RowEqualHeights,
   Scrollable,
   SearchBar,
+  SegmentedButton,
   Selectable,
   ShoppingCart,
   Skiplink,
@@ -200,6 +201,7 @@ import { cssRichContent } from "./components/rich-content/rich-content.css-templ
 import { cssRowEqualHeights } from "./components/row-equal-heights/row-equal-heights.css-template";
 import { coreScrollable } from "./components/scrollable/scrollable.core-template";
 import { cssSearchBar } from "./components/search-bar/search-bar.css-template";
+import { coreSegmentedButton } from "./components/segmented-button/segmented-button.core-template";
 import { coreSelectable } from "./components/selectable/selectable.core-template";
 import { cssShoppingCart } from "./components/shopping-cart/shopping-cart.css-template";
 import { coreSkiplink } from "./components/skiplink/skiplink.core-template";
@@ -303,6 +305,7 @@ export interface Components {
   rowEqualHeights: RowEqualHeights<TemplateResult>;
   scrollable: Scrollable<TemplateResult>;
   searchBar: SearchBar;
+  segmentedButton: SegmentedButton;
   selectable: Selectable<TemplateResult>;
   shoppingCart: ShoppingCart;
   skiplink: Skiplink;
@@ -393,6 +396,7 @@ templateContainer.add(coreProjectItem);
 templateContainer.add(coreRenvooi);
 templateContainer.add(coreResponsiveElement);
 templateContainer.add(coreScrollable);
+templateContainer.add(coreSegmentedButton);
 templateContainer.add(coreSelectable);
 templateContainer.add(coreSkiplink);
 templateContainer.add(coreSlideToggle);